Deploy services using service operator samples
Checks out the operator source and use the sample
combined with kustomize to create an instance of
the service.

A single secret gets created to have all the passwords
neede by the operators.

Kustomize is used to only set the secret and namespace.

All other changes can be done via a branch of the
operator.

For each service there is <service>_deploy_prep,
<service>_deploy and <service>_deploy_cleanup.

In addition a global `make cleanup` got added to
remove all service objects.