Add RWO,ROX,RWX to local-storage PVs
During the glance-operator development, we hit a few issues related to
the pvc stuck in Pending mode, while controller wasn't able to schedule
any pod on it.
It turned out that [1] requests the PVC with RWX, which makes sense in
a multinode environment w/ a storage class different from local-storage.
This patch adds all the potential combination needed to address any kind
of operator development.

[1] https://github.com/openstack-k8s-operators/glance-operator/blame/master/pkg/glance/pvc.go#L20

Signed-off-by: Francesco Pantano <fpantano@redhat.com>
Co-authored-by: John Fulton <fulton@redhat.com>