Support absolute path CR in *API
There are times when we want to use this project to deploy things but we
want to do so with our own CR manifest files.

For example to deploy Glance with a custom manifest to use our Ceph
cluster.

This is currently not possible, since `GLANCEAPI` is always expected to
be a relative path within the `GLANCE_REPO` repository.

This patch adds the possibility, to all the operators, of setting
absolute path using the `*API_CR` environmental variables:

- KEYSTONEAPI_CR
- MARIADB_CR
- PLACEMENTAPI_CR
- GLANCEAPI_CR
- NEUTRONAPI_CR
- CINDER_CR
- RABBITMQ_CR
- IRONIC_CR

An alternative would be to try to figure out if the `*API` environmental
variables are relative to the `*_REPO` location with something like:

GLANCEAPI_CR      = ${OPERATOR_BASE_DIR}/glance-operator/${GLANCEAPI}
ifeq ($(realpath $(GLANCEAPI_CR)),)
  GLANCEAPI_CR = $(GLANCEAPI)
endif