Add new parameters to gen-ansibleee-ssh-key.sh
In order to get rid of some ansible role in RDO, we need to be able to
reuse an existing SSH key.

This patch adds the possibility to set the existing key name, its
algorithm and key size.

The defaults values are the ones (save the default key size, sets to
4096 instead of 3072).