Move ceph container to quay.io/ceph/demo
The ceph container has been recently migrated to a new image due to a
refactoring of the repo [1].
This change aligns the Makefile to make sure we point to the right
image. In addition, no NET_DISCOVERY is available anymore, hence the
related part on the script has been removed.

[1] https://github.com/ceph/ceph-container/commit/2e0687a5147f931c5434614f3cd20f36fc6028da#diff-f6f28989a486d5db0bc66151752f59ad4a36d1b8d9f28b729609afa9f7c99014

Signed-off-by: Francesco Pantano <fpantano@redhat.com>