Use skopeo to extract CRDs in openstack_crds target
Prow only allows running in unprivileged containers, so we can't do
podman pull in there. This change uses skopeo to download the image
and then extracts the crds in the same way.

The change also adds a OPENSTACK_BUNDLE_IMG to allow taking into account
changes in CRDs introduced in a PR while running the CI jobs.