Update secret generation for dataplane-operator
The dataplane-operator now takes a Secret name that contains the private
key contents. This change updates install_yamls to match the sample
Secret name from the config/samples in dataplane-operator, and to create
the Secret with the correct format.

This change also updates the devsetup/scripts/gen-ansibleee-ssh-key.sh
script to change to the script directory intially, so that it can be run
directly from the git checkout, no matter the working directory.

Signed-off-by: James Slagle <jslagle@redhat.com>