Use DHCP reservations for edpm compute nodes
In order for the created edpm compute nodes to have known and
predictable IP addresses, use DHCP reservations. The MAC addresses are
now generated, and the IP addresses are set as reservations in the
default libvirt network from the gen-edpm-compute-node.sh script.

The IP address starts at 192.168.122.100 and the value of
$EDPM_COMPUTE_SUFFIX is added to the address. E.g., the default node
created with $EDPM_COMPUTE_SUFFIX=0 is edpm-compute-0, and it would get
the address 192.168.122.100, edpm-compute-1 would get 192.168.122.101,
etc.

The default IP address in devsetup/Makefile is updated to match the new
default, 192.168.122.100.

Signed-off-by: James Slagle <jslagle@redhat.com>