Fail if SSH_PUBLIC_KEY missing and ensure /root/.ssh exists
'make edpm_compute' runs gen-ansibleee-ssh-key.sh and then
gen-edpm-compute-node.sh. If the first script does not make
a file the second can reference then fail early to help the
user debug.

gen-edpm-compute-node.sh assumes /root/.ssh exists on the
hypervisor, but it might not. Add a condition to create
this directory if it's missing. Without this, the script
might fail while attempting to tee the public key into
root's authorized_hosts hosts file.

Signed-off-by: John Fulton <fulton@redhat.com>