Remove -t option because it has no target directory
scripts/gen-edpm-compute-node.sh added -t to its last
mv command in PR116. However, it's not a syntactically
valid use of -t and it makes the script fail with this:
mv: target '../out/edpm/id_rsa.pub' is not a directory.

This patch removes the -t to restore the original intent;
rename the file so it has the EDPM_COMPUTE_NAME appended
to it. If the -t option is going to be added back, then it
must specify a target directory. E.g. `mv -t /foo/ f1 f2`
would move files f1 and f2 to a directory named /foo/.