Remove EDPM DHCP reservation on edpm_compute_cleanup
Introduce scripts/edpm-compute-cleanup.sh and move the
Makefile entry for `edpm_compute_cleanup` to it. This
is done so that the DHCP reservation XML can be saved
to a variable and then removed with `virsh net-update`
as part of the clean up.

Running `make edpm_compute_cleanup; make edpm_compute`
more than once on the same hypervisor produces the
error below unless this patch is in place.

+ sudo virsh net-update default add-last ip-dhcp-host \
  --xml '<host mac='\''52:54:00:C1:A3:7F'\'' \
  name='\''edpm-compute-0'\'' ip='\''192.168.122.100'\''/>'\
  --config --live

error: Failed to update network default
error: Requested operation is not valid: there is an
existing dhcp host entry in network 'default' that matches
"<host mac='52:54:00:C1:A3:7F' name='edpm-compute-0'
  ip='192.168.122.100'/>"

Signed-off-by: John Fulton <fulton@redhat.com>