Remove capability to override images
The current behavior to override the image by default is can result in
an expected behavior, especially when someone wants to test a change
which updates the image name in sample CR definition.

Some components (eg. nova) do not support customizing images which is
also confusing.

In addition, the current implementation requires that we maintain
the image names in Makefile to keep these consistent with sample files
in individual repo, but this requirement is redundant.