Stop overriding service images by default
Currently install_yaml defines its own defaults about the container
images used to deploy services (eg. placement-api) but this causes
the redundant requirement that we have to maintain default images in
two layers (install_yamls and each operators).

This removes the hard-coded defaults in install_yamls and let is use
the default images defined in individual operator repos so that we can
maintain default images only in operator repos.

This also adds TODO comments for operators which do not yet have proper
interfaces for image customization, so that we can revisit these later
easily.

Finally, some lines were using tabs while the others are not. This
fixes that format inconsistency.