Run playbook deploy-edpm-growvols.yml early
This change runs the deploy-edpm-growvols.yml playbook as early as
practical, so that the volumes are grown to the space of the disk
before any other activity does any writes.

The edpm_growvols role is a noop on target hosts which do not have the
growvols utility installed on them. This playbook needs to be in place
before any switch to a custom EDPM image.