Support RHEL8 Hypervisors without --firstboot-command
A recent patch [1] added a call to `virt-customize
--firstboot-command` but that option isn't supported
in the `virt-customize` which ships with RHEL8.

This patch switches from --firstboot-command to
--firstboot  which is supported on the RHEL8 version
of virt-customize (1.44 as of now). This has the same
effect but let's us continue to support RHEL8.

Also, the older version of virt-copy-out is not
using the exported VIRSH_DEFAULT_CONNECT_URI so
pass it with a `-c`.

[1] 0cb376b52e3f0e299fb40dd1f5c0eb8ff51cb492

Signed-off-by: John Fulton <fulton@redhat.com>