Generalize the cleanup strategy
When openstack_cleanup is run, it might happen that we have more than one
csv/subs/catalogsource in the same namespace. This is the case of
openstack-operator and manila-operator: even though manila is deployed
by meta operator, the OLM dependency model installs separated resources.
This patch introduces a generalized way to cleanup the namepace, where
for each operator that has a dedicated subscription, the associated
resources are deleted.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>