Fix defaulting of the mariadb service image
We use the "unused" string in the Makefile to indicate that no
kustomization is needed for the given image field. For the mariadb
service image there was an extra set of double quotes " was added to the
string resulting in that the mariadb service image is always kustomized
to the "unused" string.