Fix container image override for glance deployment
The glances CR has three containerImage key in its spec. Currently
only the top-level one is overridden and actual glance-api services are
always deployed using the default image. This fixes that and ensures
all container images are overridden.

This also updates the kustomize script so that we can override multiple
images.