Allow to NOT rm generated content
It may happen we'd want to keep the generated content from previous run
or integrate external content, generated before hitting a specific step.

This is especially true for content related to kustomize: we may get an
external tool (like ci-framework) generating some content in the
DEPLOY_DIR; with the call to _cleanup tasks as preparation of _prep
tasks, that generated content would then be lost, leading to an unwanted
behavior.

This patch allows to override the "rm -Rf" usually called, by providing
a new parameter CLEANUP_DIR_CMD.
It's also exported as an environment variable, therefore it should allow
nested scripts to benefit of it.