Allow git clone commands to take additional parameters
Till now git clones assumed that the source code to be used
can be defined by the repo and branch tuple, but if the code
comes from a PR we need to add the reference of the PR into
the equation.
To avoid overcomplicating git clone management this change
proposes to allow external customization of the git clone
command, this way, install_yamls users can, for example,
clone for a local repo already pointing to the desidered
reference and speeding up the clone by providing the "-l"
option.