Install repo-setup without installing git or pip
This switches the repo-setup install to download a github tarball
and using the pip bundled with python to install repo-setup in a venv.

This means no packages need to be installed before the repos are set
up.

The script is run with the root user, so sudo calls have been removed
also.