Customize the amount of storage associated to Ceph
We recently hit issues in CI because this pod is evicted with the
following message:

'''
Usage of EmptyDir volume "data" exceeds the limit "500Mi".
'''

This patch exposed the DATA_SIZE parameter that can be used to
customize the amount of space reserved for the toy ceph cluster
(especially for CI purposes).

Signed-off-by: Francesco Pantano <fpantano@redhat.com>