Apply OVNController CRD if it's present
The new OVNController CRD is about to be introduced in ovn-operator
(it's the old OVS CRD from ovs-operator - we are merging ovn- and ovs-
into one operator).

This change will make sure that the new manifest from ovn-operator repo
is applied, if present. This is needed for CI of ovn-operator repo to
pass.