Add deps for dataplane-operator kuttl tests
The SSH key secret needs to exist for the dataplane-operator kuttl tests
to pass. Call the script that creates the secret before running tests.

The initial dataplane-operator services also need to be created from the
config/services directory in the operator.

The tests also require the openstack-ansibleee-operator. The ansibleee
Makefile target is added as a dependency for the tests.

Signed-off-by: James Slagle <jslagle@redhat.com>