Remove ssh-keyscan from gen-edpm-compute-node script
The ssh-keyscan will fail if there is no info on known_hosts for the ip
address 192.168.122.1 and so the virt-customize will command fail to
run, specially when you're trying to deploy locally in your machine
instead of using nested vms.
Also, set the proper permission for the firstboot script, since one of
my tests, when you run virt-customize, it fails without the proper
executable permissions.