edpm_compute switch from centos to edpm-hardened-uefi
This switches a crc environment to the edpm-hardened-uefi.qcow2 image
which is distributed in a container image (build defined in [1]).

Automation will rebuild the edpm-hardened-uefi image as required, so the
container image
quay.io/podified-antelope-centos9/edpm-hardened-uefi:current-podified
will remain up to date.

In the hardened image, /tmp is mounted as noexec, so repo setup has been
moved to /var/tmp

[1] https://github.com/openstack-k8s-operators/edpm-image-builder/