Customize the updated image path for custom ansible image
dataplane-operator commit 17e87a135fc6671129c921560e26fd7ecf259113 moved
the OpenStackAnsibleEEImage field from the node/role to the service so
that the image can be customized per service.

This patch updates install_yamls to kustomize the correct path on the
service when a custom ${DATAPLANE_RUNNER_IMG} is set.

Signed-off-by: James Slagle <jslagle@redhat.com>