Added vars to update container urls for EDPM
In periodic integration lines, We build and push
the openstack services containers to a different regitry and with
different hash.

In order to deploy EDPM in periodic integration lines, we need to update
the openstack services containers with registry url and hash from
periodic line to test the same.

This patch adds two vars to update the registry url with namespace
and proper container hash for edpm and baremetal job.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>