Make devsetup/cifwm_prepare not fail when rerun
Currently if one tries to rerun cifwm_prepare it fails
on git cloning attempt since the directory is already in place.

Example of failure:
> + make cifmw_prepare
> git clone https://github.com/openstack-k8s-operators/ci-framework ci-framework
> fatal: destination path 'ci-framework' already exists and is not an empty directory.
> make: *** [Makefile:180: cifmw_prepare] Error 128

As that is not expected from make target, make it so that
this target depends on the directory existing target,
which itself calls the git clone for its creation.