Fix MON_IP in the Ceph pod for multinode OCP CI
The Ceph pod is supposed to be single node, single pod, single
container. If MON_IP is not passed and computed by the "get_pod_ip"
function, in a multinode context it might return a list of Internal IP
addresses. This patch represents an attempt to only support a single IP
address from the list and assign it to the mon spawned as first process
during the cluster bootstrap.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>