Check operator status to ensure operator is installed
... instead of checking deployment status. This is more robust way to
ensure all resources including CRDs are installed. Also, ensure
the required operator is ready when creating a CR.

With this change we'll remove openstack_crds target from kuttl targets.
The problem with using openstack_crds in kuttl targets is that the CRDs
in the openstack operator bundle image can contain old CRD which does
not match the actual CRD version expected by the operator image we are
deploying.