Ensure we have a clean exit condition in the wait loop
Some `while` loop were added in the past, in order to mostly wait for a
resource to be created.

It may happen the resource never gets created or to the expected state.
In such condition, we end in an infinite loop.

Leveraging the `timeout` app, we may ensure we exit the loop sooner or
later.