Allow to set MTU for networks
It may happen we can't configure the underlying network with an MTU
higher to 1500. In such case, network isolation will not be possible.

Allowing to configure isolated network MTU should allow to deploy on any
kind of environment, and actually be closer to actual production use
where customer can't or don't want to enable jumbo frames.