Allow passing the Ansible user as a parameter of edpm_deploy
With the new CI that uses multiple preprovisioned nodes we need
to be able to uses non-default users for connection to each compute
from Ansible. This commit add a new optional parameter
DATAPLANE_ANSIBLE_USER that, if given, will add a new entry to the
already existing kustomization of the dataplane resource to replace
the ansibleUser variable.