Addional Ceph mon related config for multinode OCP
When HOSTNETWORK is false and we run a multinode cluster, we want to use
the podIP to run the mon, so that it will be accessible in the k8s
overlay from all the ctlplane pods using the openshift-sdn.
This patch adds the additional MON_CONF to the main Ceph CR with the
required bits to get the POD IP address before running the actual mon
process.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>