Limit affinity of local storage pv
When a pod claims a persistent volume that is configured to store
data locally to a worker node, it should not be allowed to restart
on another worker node or it will not see the expected data.

Change the way we generate the pv to only allow a pod to restart
locally once it claimed a pv, by generating pv for each worker
node in the cluster.