Remove unnecessary kustomize op to nonexistent ansibleVars
Update scripts/gen-edpm-kustomize.sh so that the generated
kustomization.yaml does not try to remove a nonexistent key.

Now that 38c4889 has merged there will not be any ansibleVars
string under /spec/nodes/edpm-compute-0. So, in /spec/nodes/,
after edpm-compute-0 is copied to edpm-compute-${INDEX}, there
will be no need to delete it. Also, the attempt to delete it
results in the following error when EDPM_SINGLE_NODE is false
and EDPM_TOTAL_NODES is >2.

"Unable to remove nonexistent key: ansibleVars: missing value"

Signed-off-by: John Fulton <fulton@redhat.com>
Jira: OSP-27127