Modify nncp to allow interface/network/vlan customization
Adopting a multicontroller OSP 17.1 wallaby deployment means
modifying a lot of the hardcoded values in nncp stage.
In order to make things easier the Interfaces, Networks and vlans
are parameterized.
As well as adding a new variables section called MCD (multi cluster
deployment) with the correct values to adopt a multi controller
wallaby environment.