use tripleo-repos form current-tripleo on stable wallaby
This change updates the location where the python3-tripleo-repos
rpm is installed from to use the promoted content from the stable
wallaby line.

This corrects the fact the rpm is nolonger avaiable at
https://trunk.rdoproject.org/centos9/component/tripleo/current/
currenlty, ignoring the tempory issue this should also be more
stable as we are installing stable wallaby anyway and tripleo
master is nolonger used for development. stable/wallaby is the
maintained branch so we should align to that.