Update for using OpenStackDataPlaneDeployment
Deploying of the NodeSet is triggered with an
OpenStackDataPlaneDeployment CR instead of the deployStrategy field. The
OpenStackDataPlaneDeployment CR, configured with
$OPENSTACK_DATAPLANEDEPLOYMENT, and identified by
$DATAPLANE_DEPLOYMENT_CR, is copied into the $DEPLOY_DIR for
dataplane-operator resources.

It will be created during the edpm_deploy target when kustomize is run
against the $DEPLOY_DIR, triggering the deployment at the same step as
before.

The edpm kustomize scripts are updated to not set any fields under
deployStrategy, as that is no longer used.

See: https://github.com/openstack-k8s-operators/dataplane-operator/pull/378
Signed-off-by: James Slagle <jslagle@redhat.com>