devsetup - Standalone ironic resolv apps-crc.testing
To enable network connection to the virtual RedFish BMC
running in crc configure the primary nameserver to use
the crc address (192.168.130.11) when compute driver is
ironic.

For this to work we need to set up routing for the
libvirt networks, which is done in PR# 564 and a host
route must be added to the edpm standalone node. The
route can be added in EDPM_COMPUTE_ADDITIONAL_NETWORKS
as shown in below.

```bash
cat << EOF > /tmp/addtional_nets.json
[
  {
    "type": "network",
    "name": "crc-bmaas",
    "standalone_config": {
      "type": "linux_bridge",
      "name": "baremetal",
      "mtu": 1500,
      "ip_subnet": "172.20.1.0/24",
      "allocation_pools": [
        {
          "start": "172.20.1.4",
          "end": "172.20.1.250"
        }
      ],
      "host_routes": [
        {
	  "destination": "192.168.130.0/24",
          "nexthop": "172.20.1.1"
	}
      ]
    }
  }
]
EOF
export EDPM_COMPUTE_ADDITIONAL_NETWORKS=$(jq -c . /tmp/addtional_nets.json)
```

Signed-off-by: Harald Jensås <hjensas@redhat.com>