devsetup - handle if xmllint does not return an IP
Fixes issue from 071252c37b7bdfe5ae8eb88bdef8d77be300898b

On a RHEL8 hypervisor a firstboot script contained the full
XML output of `virsh net-dumpxml` instead of an IP address.
If a valid IP address is not returned from a call to virsh
net-dumpxml and xmllint, then set EDPM_COMPUTE_NETWORK_IP
to a default of 192.168.122.1 as a fallback.

Signed-off-by: John Fulton <fulton@redhat.com>