bmaas - log virtual baremetal console to file
Set up the virtual baremetal nodes created to log to a file.

Inject a libvirt hook that will archive the console log's when
a vm is powered off. Logs are archived in two formats, the
original log file including ANSI control codes and a _no_ansi_
log with ANSI codes removed.

Signed-off-by: Harald Jensås <hjensas@redhat.com>