Add support for setting image_prefix in EDPM openstack services images
In downstream, the openstack services container images start with
osp18-openstack. In upstream, it is openstack.

In order to use the correct image downstream in CI, we need to update
the image prefix.

https://github.com/openstack-k8s-operators/dataplane-operator/pull/489
already adds the support of image_prefix in dataplane crs. This pr
adds a var in install_yamls to update the image_prefix in dataplane cr.

Depends-On: https://github.com/openstack-k8s-operators/dataplane-operator/pull/489

Signed-off-by: Chandan Kumar <raukadah@gmail.com>