dnsmasq in CentOS is old, compile from git ...
The dnsmasq version in CentOS 9 Stream does not support
the --filter-A and --filter-AAAA functionality.

Detect this, and combile dnsmasq from git in this case.
Install under the workdir which is not in the path so that
only the systemd services created for the ipv6nat64 lab use
the compiled dnsmasq.

Signed-off-by: Harald Jensås <hjensas@redhat.com>