network-isolation using custom libvirt network
Add conditions based on using the "default" libvirt network
or a custom libvirt network for network-isolation.

When using the "default" network, use addressing in 192.168.122.0/24,
when not using the "default" network use 172.16.1.0/24. (This range
matches the default for the libvirt network created by devsetup/Makefile
target ``network_isolation_bridge``.

To use the custom libvirt network:
  make network_isolation_bridge
  export NETWORK_ISOLATION_USE_DEFAULT_NETWORK=false
  < .. commands as usual ..>

Signed-off-by: Harald Jensås <hjensas@redhat.com>