Ensure early RNG entropy, avoid FIPS kernel panic
Currently there are frequent kernel panics in CI caused by "Jitter RNG
permanent health test failure". This is likely due to insufficient
entropy early in the boot.

This panic occurs at ~500ms into boot, and the default period for
consuming from source is 1000ms.

This is an attempt to improve CI stability by reducing the period to
250ms.

Jira: OSPRH-1967