Add the ability to deploy multiple Ceph Pods
We recently introduced a feature that allows to deploy an arbitrary numbers
of GlanceAPI [1].
To ease the test and to be closer to an edge scenario simulation, an
approach would be to have the ability to deploy multiple Ceph pods:
they bring different secrets that can be propagated to a subset of
ctlplane components.
This patch introduces the new 'CEPH_CLUSTERS' variable that is used
within the bash script to (eventually) deploy multiple Ceph Pods.

[1] https://github.com/openstack-k8s-operators/glance-operator/pull/384

Signed-off-by: Francesco Pantano <fpantano@redhat.com>