Allow sushy emulator to use openstack
This change allows sushy-tools to be started with openstack as the
driver instead of libvirt. With the current default values there is no
change from the current libvirt behaviour.

With this change, a configured sushy-emulator pod can be started which
is managing openstack instances instead of libvirt VMs.

Further feature work is required on sushy-tools for this to be actually
useful. This is the beginning of reworking of baremetal jobs to become
multinode to improve reliability.