Configure OVN Controller to act as gateway
By default we do not configure nicMappings for
ovnController CR as it requires dedicated nic
with external connectivity on the worker nodes.

With network isolation we already setup an additional
nic on the worker nodes and as that is already used by
MetalLB and net-attach-definitions it cannot be
shared by OVN Controllers as it uses host-device
plugin and requires a dedicated NIC.

In order to use same interface for ovn controller,
a linux bridge is created with NNCP_INTERFACE as one
of it's port. Utilizing this bridge a net-attach-def of
bridge type is created for usage by the OVN Controllers.

Related-Issue: OSPRH-649