Add make target to create edpm-networker node(s)
In devsetup's makefile there is now target to create and clean
edpm-networker node(s). It is very similar to edpm-compute but it has
different name of the VM, just to be able to distinguish easier
networker and compute in the dev env.

It also changes names of the gen-edpm-compute-node and
edpm-compute-cleanup scripts to make them more generic and clear that
same scripts are used also for the edpm-networker nodes now.
The same rename is also done for the bgp node scripts.