Fix usage of multiple eval calls in a make target
eval get's executed first of all irrespective of it's
position when running a make target.

So multiple calls to same function even with different arguments
in a make target do not work as expected as only the last call
overrides others. This patch drops redundant calls and pass
required arguments explicitly wherever needed.